+++ local/questionbulkupdate/lib.php
@@ -125,9 +125,10 @@ function local_questionbulkupdate_extend_settings_navigation(navigation_node $na
     );
 
     //ecastro ULPGC
+    /*
     $url = new moodle_url('/local/questionbulkupdate/bulkupdate.php', array('courseid' => $context->instanceid));
     $newnode = navigation_node::create(get_string('navandheader', 'local_questionbulkupdate'),
                 $url, navigation_node::TYPE_SETTING, null, 'cquestionbulkupdate');
-    $parentnode->add_node($newnode, 'contentbank');  
-    
+    $parentnode->add_node($newnode, 'roleassign');  
+    */
 }