+++ admin/tool/redis/CHANGES.md
@@ -4,6 +4,11 @@ moodle-tool_redis
 Changes
 -------
 
+### v4.1-r2
+
+* 2023-05-03 - Bugfix: Get rid of a debugging message in redis-stats library on PHP 8.1.
+* 2023-04-30 - Tests: Updated Moodle Plugin CI to use PHP 8.1 and Postgres 13 from Moodle 4.1 on.
+
 ### v4.1-r1
 
 * 2023-04-30 - Update redis-stats library to latest version.