+++ theme/boost_union/CHANGES.md
@@ -4,8 +4,10 @@ moodle-theme_boost_union
 Changes
 -------
 
-### Unreleased
+### v4.1-r7
 
+* 2023-04-20 - Feature: Course category breadcrumbs in course header, solves #284.
+* 2023-04-02 - Improvement: Add SCSS code to improve the block regions presentation in our companion plugin Dash Pro, solves #291.
 * 2023-04-17 - Bugfix: Setting activityiconcolorcommunication was not processed anymore, solves #279.
 * 2023-04-13 - Bugfix: When there is no edit switch, as on the assignment grading page, scrollfix should not kick in, solves #276.
 * 2023-04-12 - Feature: Allow right-side blocks drawer of site home to be extended by default #169.